Renamed dialogic to ravestate.
