--- modules/dialogic_tts_watson/__init__.py
@@ -1,5 +1,0 @@
-
-from dialogic import registry
-from dialogic_tts_watson import states
-
-registry.register(name="tts_watson", states=(states.audio_talk,))
