+++ modules/ravestate_tts_watson/__init__.py
@@ -1,0 +1,5 @@
+
+from ravestate import registry
+from ravestate_tts_watson import states
+
+registry.register(name="tts_watson", states=(states.audio_talk,))
