+++ modules/ravestate_ui/templates/index.html
