+++ test.py
@@ -3,10 +3,10 @@ import sys
 import os
 sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/modules")
 
-from dialogic.context import Context
-from dialogic.state import state
-from dialogic import registry
-from dialogic_hmi import service
+from ravestate.context import Context
+from ravestate.state import state
+from ravestate import registry
+from ravestate_ui import service
 
 
 @state(triggers=":startup", write="rawio:out")
@@ -25,13 +25,12 @@ def face_recognized(ctx):
         ctx["rawio:out"] = "I see you, {}!".format(ctx["facerec:face"])
 
 
-registry.register(name="hi", states=(hello_world, generic_answer))
+registry.register(name="hi", states=(hello_world, generic_answer, face_recognized))
 
 ctx = Context()
-ctx.add_module("dialogic_rawio")
-ctx.add_module("dialogic_consoleio")
-ctx.add_module("dialogic_tts_watson")
-#ctx.add_module("dialogic_facerec")
+ctx.add_module("ravestate_conio")
+# ctx.add_module("ravestate_tts_watson")
+# ctx.add_module("ravestate_facerec")
 ctx.add_module("hi")
 ctx.run()
 service.advertise(ctx=ctx)
